Time: 0 ms (100%), Space: 95.1 MB (78.85%) - LeetHub

+++ 0004-median-of-two-sorted-arrays/0004-median-of-two-sorted-arrays.cpp
@@ -19,7 +19,7 @@ public:
             if(mid2<n2) r2=nums2[mid2];
             if(l1<=r2&&l2<=r1){
                 if(n%2==1) return (double)max(l1,l2);
-            else return (double)((max(l1,l2)+min(r1,r2))/2.0);}
+            else return ((max(l1,l2)+min(r1,r2))/2.0);}
             else if(l1>r2)
             high=mid1-1;
             else
